Almost there with pooled bullets
So the bullet pointer implementation still doesn't work as I think it
should, however I have narrowed the problem down to oop I believe.  I
created an array containing a bunch of bullets, or so I thought, when
actually I believe I created an array of pointer all pointing to the
same bullet.  Once I get to school in a few hour I'll have a second pair
of eyes look at the issue to see if I can get it sorted quickly.